Merge in preparation for production.

+++ commands/roles.js
@@ -1,12 +1,13 @@
 const Discord = require('discord.js');
 const Settings = require('../models/settings.js');
 const { noPerms, maintenanceMode } = require('../utils/errors.js');
-const { orange, owner } = require('../config.json');
+const { embedColor, owner } = settings;
 
 exports.run = async (client, message, args) => {
     
     let perms = message.guild.me.permissions;
     if (!perms.has('MANAGE_MESSAGES')) return message.channel.send('I can\'t delete messages! Make sure I have this permission: Manage Messages`').then(msg => msg.delete(5000));
+    if (!perms.has('EMBED_LINKS')) return message.channel.send('I can\'t embed links! Make sure I have this permission: Embed Links`').then(msg => msg.delete(5000));
 
     await message.delete().catch(O_o => {});
 
@@ -15,58 +16,50 @@ exports.run = async (client, message, args) => {
         return maintenanceMode(message.channel);
     }
 
-    Settings.findOne({
-        guildID: message.guild.id,
-    }, (err, res) => {
-        if (err) return console.log(err);
-
-        const roles = res.staffRoles;
-
-        const staffRoles = roles.map(el => {
-            return '<@&' + el.role + '>';
-        });
-
-        let adminPerms;
-        let admins = [];
-        message.guild.members.forEach(collected => {
-            if (collected.hasPermission('MANAGE_GUILD') && !collected.user.bot) {
-                
-                admins.push(collected.id);
-
-                return adminPerms = admins.map(el => {
-                    return '<@' + el + '>';
-                });
-            }
-        });
-
-        if (!admins.includes(message.member.id)) return noPerms(message, 'MANAGE_GUILD');
-
-        let embed = new Discord.RichEmbed()
-            .setColor(orange)
-            .addField('Admins', adminPerms.join('\n'));
-
-        for (let i = 0; i < 1; i++) {
-            try {
-                embed.addField('Staff Roles', staffRoles.join('\n'));
-            } catch (err) {
-                break;
-            }
-        }
-
-        let perms = message.guild.me.permissions;
-        if (!perms.has('EMBED_LINKS')) return message.channel.send('I can\'t embed links! Make sure I have this permission: Embed Links`').then(msg => msg.delete(5000));
-
-        return message.channel.send(embed);
+    let gSettings = await Settings.findOne({ guildID: message.guild.id }).catch(err => {
+        console.log(err);
+        return message.channel.send(`Error querying the database for this guild's information: **${err.message}**.`);
     });
-};
 
-exports.conf = {
-    aliases: ['viewroles', 'viewrole', 'staffroles'],
-    status: 'true'
+    const roles = gSettings.staffRoles;
+
+    const staffRoles = roles.map(el => {
+        return '<@&' + el.role + '>';
+    });
+
+    let adminPerms;
+    let admins = [];
+    message.guild.members.forEach(collected => {
+        if (collected.hasPermission('MANAGE_GUILD') && !collected.user.bot) {
+
+            admins.push(collected.id);
+
+            return adminPerms = admins.map(el => {
+                return '<@' + el + '>';
+            });
+        }
+    });
+
+    if (!admins.includes(message.member.id)) return noPerms(message, 'MANAGE_GUILD');
+
+    let embed = new Discord.RichEmbed()
+        .setColor(embedColor)
+        .addField('Admins', adminPerms.join('\n'));
+
+    for (let i = 0; i < 1; i++) {
+        try {
+            embed.addField('Staff Roles', staffRoles.join('\n'));
+        } catch (err) {
+            break;
+        }
+    }
+
+    message.channel.send(embed);
 };
 
 exports.help = {
     name: 'roles',
+    aliases: ['viewroles', 'viewrole', 'staffroles'],
     description: 'View the current staff roles for the bot',
     usage: 'roles'
 };
